+++ versions.tf
@@ -1,17 +1,17 @@
 terraform {
-  required_version = ">= 1.7.0"
+  required_version = "~> 1.9"
   required_providers {
     aws = {
       source  = "hashicorp/aws"
-      version = ">= 5.38.0"
+      version = "~> 5.38"
     }
     helm = {
       source  = "hashicorp/helm"
-      version = ">= 2.13.2"
+      version = "~> 2.13"
     }
     kubernetes = {
       source  = "hashicorp/kubernetes"
-      version = ">= 2.30.0"
+      version = "~> 2.30"
     }
   }
 }