+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2024-10-20
+### Updated
+- Version EKS module
+- Version Karpenter module
+- pre-commit hooks config
+### Added
+- Secret for OIDC Auth
+### Changed
+- automated_prune and automated_selfHeal now are configurable.
+
 
 ## [1.0.0] - 2024-07-08
 ### Added